% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{decimals_only}
\alias{decimals_only}
\title{Force printing a specified number of decimals and leave out a leading
zero}
\usage{
decimals_only(x, n_digits)
}
\arguments{
\item{x}{the value to be printed}

\item{n_digits}{how many decimals are to be printed}
}
\value{
The number in the required format
}
\description{
Force printing a specified number of decimals and leave out a leading
zero
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
