% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_decimals_specified.R
\name{force_decimals}
\alias{force_decimals}
\title{Force printing a specified number of decimals for a number}
\usage{
force_decimals(x, decimals = 2)
}
\arguments{
\item{x}{the numeric values to be printed}

\item{decimals}{how many decimals are to be printed. Defaults to 2.}
}
\value{
The number in the required format
}
\description{
Force printing a specified number of decimals for a number
}
\examples{

force_decimals(c(1.23456, 0.873, 2.3456))
force_decimals(c(1.23456, 0.873, 2.3456), 3)

}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
