% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{force_or_cut}
\alias{force_or_cut}
\title{Print a number}
\usage{
force_or_cut(x, n_digits)
}
\arguments{
\item{x}{A number}

\item{n_digits}{The number of digits that should be printed if x is a decimal number}
}
\value{
The number in the required format
}
\description{
Print a number
}
\details{
If x integer, only the integer is printed, if x is a decimal
    number, the decimals are printed
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
