% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{format_p}
\alias{format_p}
\title{Format a p-value according to APA standards}
\usage{
format_p(pvalue, decimals)
}
\arguments{
\item{pvalue}{The p-value}

\item{decimals}{The number of decimals to be printed}
}
\value{
A string representation of the p value to be used in Rmarkdown
  documents.
}
\description{
Format a p-value according to APA standards
}
