% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_anova.R
\name{print_anova}
\alias{print_anova}
\title{Print the results of an `afex` ANOVA object}
\usage{
print_anova(afex_object, es, font = "italic", decimals = 2,
  decimals_p = 3)
}
\arguments{
\item{afex_object}{an object returned by `afex`s ANOVA functions}

\item{es}{A String "pes" or "ges"; which effect size measure was
used?}

\item{font}{should the effect size symbol eta be printed in "italic"
or "nonitalic". See details.}

\item{decimals}{how many decimals should be printed}

\item{decimals_p}{how many decimals should be printed for the p-value
(defaults to 3)}
}
\description{
Print the results of an `afex` ANOVA object
}
\details{
According to APA style, the eta symbol should be printed non-italic,
but the standard Latex \eta symbol is italic. To print a 
non-italic eta, use font = "nonitalic". However, this option 
requires that you load the package `upgreek` in the YAML header of 
your R markdown document. To this end, use the following: 

header-includes:
  -\usepackage{upgreek}
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
