% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print_chi2}
\alias{print_chi2}
\title{Printing the results of a chi-square test}
\usage{
print_chi2(tab = NULL, chi2.object = NULL, es = TRUE,
  correct = FALSE, decimals = 2, decimals_p = 3)
}
\arguments{
\item{tab}{A contingency table}

\item{chi2.object}{an object that is returned by `chisq.test` (do 
not pass if argument `tab` is passed).}

\item{es}{Boolean. Should the phi coefficient be printed (only makes
sense for 2x2 contingency tables!)}

\item{correct}{Boolean. Apply a continuity correction? See `?chisq.test`}

\item{decimals}{How many decimals should be printed}

\item{decimals_p}{How many decimals should be printed for the p-value
(defaults to 3)}
}
\value{
A string describing the results of the chi-square test to be 
  printed in Rmarkdown documents.
}
\description{
Printing the results of a chi-square test
}
\details{
The returned effect size (the phi-coefficient) only makes
    sense for 2x2 contingency tables.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
