% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print_cortest}
\alias{print_cortest}
\title{Printing the results of a significance test for a correlation
coefficient}
\usage{
print_cortest(cor_object, decimals = 2, decimals_p = 3)
}
\arguments{
\item{cor_object}{An object returned by `cor.test`}

\item{decimals}{how many decimals should be printed}

\item{decimals_p}{how many decimals should be printed for the p-value
(defaults to 3)}
}
\value{
A string describing the test
}
\description{
Printing the results of a significance test for a correlation
coefficient
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
