% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ttest.R
\name{print_ttest}
\alias{print_ttest}
\title{Print the results of a t-test}
\usage{
print_ttest(t_object, d_object, decimals = 2, decimals_p = 3,
  paired = FALSE)
}
\arguments{
\item{t_object}{an object returned by `afex`s ANOVA functions}

\item{d_object}{An object returned by `effsize::cohen.d`}

\item{decimals}{how many decimals should be printed}

\item{decimals_p}{how many decimals should be printed for the p-value
(defaults to 3)}

\item{paired}{Logical vector of length 1. Was the t-test a
within-subjects comparison? Determines whether Cohen's d is
printed as d_z when TRUE. Defaults to FALSE.}
}
\description{
Print the results of a t-test
}
\details{
Note that the R package `papaja` contains functionality for
    printing many different statistical tests. This is just for my
    personal use that I have used it in recent writeups.
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
