% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prmisc-package.R
\docType{package}
\name{prmisc-package}
\alias{prmisc-package}
\title{The prmisc Package}
\description{
Miscellaneous printing.
}
\details{
\tabular{ll}{
Package: \tab prmisc\cr
Type: \tab Package\cr
Version: \tab 0.01-0\cr
Date: \tab 2018-05-09\cr
Depends: \tab R (>= 3.0.0)\cr
Encoding: \tab UTF-8\cr
License: \tab Public domain\cr
LazyLoad: \tab yes\cr
URL: \tab https://github.com/m-Py/prmisc\cr
}

Miscellaneous printing of stat results in Rmarkdown
}
\author{
Martin Papenberg
}
\keyword{package}
