% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_only_decimals.R
\name{decimals_only}
\alias{decimals_only}
\title{Force printing a specified number of decimals and leave out a leading
zero}
\usage{
decimals_only(x, decimals = 2)
}
\arguments{
\item{x}{the values to be printed}

\item{decimals}{how many decimals are to be printed. Defaults to 2.}
}
\value{
The number in the required format
}
\description{
Force printing a specified number of decimals and leave out a leading
zero
}
\examples{

decimals_only(c(0.23456, 0.873, 0.3456), decimals = 3)

}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
