% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_only_decimals.R
\name{decimals_only}
\alias{decimals_only}
\title{Printing a specified number of decimals and ignore leading zeros}
\usage{
decimals_only(x, decimals = 2, decimals1 = FALSE)
}
\arguments{
\item{x}{the values to be printed}

\item{decimals}{how many decimals are to be printed. Defaults to 2.}

\item{decimals1}{Boolean - should a value of exactly 1 be converted 
to 1.00. Defaults to FALSE.}
}
\value{
The number in the required format
}
\description{
Printing a specified number of decimals and ignore leading zeros
}
\examples{

decimals_only(c(0.23456, 0.873, 0.3456), decimals = 3)

}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
