% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{force_decimals}
\alias{force_decimals}
\title{Force printing a specified number of decimals for a number}
\usage{
force_decimals(x, decimals)
}
\arguments{
\item{x}{the values to be printed}

\item{decimals}{how many decimals are to be printed}
}
\value{
The number in the required format
}
\description{
Force printing a specified number of decimals for a number
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
