% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{force_or_cut}
\alias{force_or_cut}
\title{Print a number having a specified number of digits or as integer}
\usage{
force_or_cut(x, decimals)
}
\arguments{
\item{x}{A vector of numbers}

\item{decimals}{The number of digits that should be printed if x is a
decimal number}
}
\value{
The number in the required format
}
\description{
Print a number having a specified number of digits or as integer
}
\details{
If x integer, only the integer is printed, if x is a decimal
    number, the decimals are printed
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
