% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print_anova}
\alias{print_anova}
\title{Print the results of an `afex` ANOVA object}
\usage{
print_anova(afex_object, row, es, font = "nonitalic", decimals = 2,
  decimals_p = 3)
}
\arguments{
\item{afex_object}{an object returned by `afex`s ANOVA functions}

\item{row}{the number of the row of the effect in the table view of
the `afex_object`}

\item{es}{A String "pes" or "ges"; which effect size measure was
used?}

\item{font}{should the effect size symbol eta be printed in "italic"
or "nonitalic"}

\item{decimals}{how many decimals should be printed}

\item{decimals_p}{how many decimals should be printed for the p-value
(defaults to 3)}
}
\description{
Print the results of an `afex` ANOVA object
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
