% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print_chi2}
\alias{print_chi2}
\title{Printing the results of a chi-square test}
\usage{
print_chi2(tab = NULL, chi2.object = NULL, es = TRUE, correct = FALSE,
  decimals = 2, decimals_p = 3)
}
\arguments{
\item{tab}{A contingency table. Do not combine with argument
`chi2.object`.}

\item{chi2.object}{an object that is returned by `chisq.test`. Do not
combine with argument `tab`. Can also handle objects returned by 
`spgs::chisq.unif.test`.}

\item{es}{Boolean. Should the phi coefficient be printed. This argument
only has an effect if `tab` is passed as a 2x2 contingency
table. Does not have an effect if `chi2.object` is passed.}

\item{correct}{Boolean. Apply a continuity correction? See
`?chisq.test`}

\item{decimals}{How many decimals should be printed}

\item{decimals_p}{How many decimals should be printed for the p-value
(defaults to 3)}
}
\value{
A string describing the results of the chi-square test to be 
  printed in Rmarkdown documents.
}
\description{
Printing the results of a chi-square test
}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
