% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ttest.R
\name{print_ttest}
\alias{print_ttest}
\title{Print the results of a t-test}
\usage{
print_ttest(t_object, d_object, decimals = 2, decimals_p = 3,
  paired = FALSE)
}
\arguments{
\item{t_object}{An object returned by \code{\link{t.test}}}

\item{d_object}{An object returned by \code{\link{effsize::cohen.d}}`}

\item{decimals}{How many decimals should be printed for r and t-values.}

\item{decimals_p}{How many decimals should be printed for the p-value
(defaults to 3)}

\item{paired}{Logical vector of length 1. Was the t-test a
within-subjects comparison? Determines whether Cohen's d is
printed as d_z when TRUE. Defaults to FALSE.}
}
\description{
Print the results of a t-test
}
\details{
To use this function, you need to install the R package effsize and 
compute a Cohen's d object; pass this object as the second argument.
}
\examples{

ttest <- t.test(1:10, y = c(7:20), var.equal = TRUE)
library("effsize") # for Cohen's d
cohend <- cohen.d(1:10, c(7:20))
print_ttest(ttest, cohend) # include this call in Rmd inline code

}
\author{
Martin Papenberg \email{martin.papenberg@hhu.de}
}
